Restore autoconfed kernel route for gw if missing (#3190)
The Antrea Agent has logic to periodically sync gateway routes and to
ensure that any route deleted by mistake is restored. However this logic
did not include the subnet route automatically configured by the kernel
when an IP address is configured on the gateway interface. We update the
logic to include this route.

Fixes #3187

Signed-off-by: Antonin Bas <abas@vmware.com>