Fix race condition in TestAggregatorCollectListGet unit test (#3305)
Before this fix it was possible for the Run method to start processing
summaries before all ADD events had been processed by the registered
event handlers. When that happened, the summary was discarded by
doCollect. The reason is that there is no guarantee that all event
handlers have completed by the time the cache reports that it has
synced.

Another issue was that len(a.dataCh) being 0 is not a good enough signal
to detect that summaries have been processed. The Run method may still
be executing doCollect for the last summary. To avoid this issue, we
close stopCh and wait for Run to return.

Fixes #3300

Signed-off-by: Antonin Bas <abas@vmware.com>